renames files and reverts index.jsx changes. also uses git mv to push files with new names
