+++ src/components/flexView/SeriesTable.jsx
