+++ src/components/flexView/StudyTable.jsx
